Move protocol implementation into its own file
